# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Flexfence Backend API
  version: 0.1.0
  description: |
    MVP attendance API for Flexfence.
servers:
  - url: http://localhost:8080
...
paths:
...
  /v1/me/event-joins:
    get:
      tags: [Events]
      summary: List events the signed-in attendee has joined
      operationId: listMyEventJoins
      security:
        - UserBearer: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: q
          in: query
          description: Search event title or description
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: join_source
          in: query
          schema:
            type: string
        - name: start_from
          in: query
          schema:
            type: string
            format: date-time
        - name: start_to
          in: query
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: Join history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListUserEventJoinsResponse"
        default:
          $ref: "#/components/responses/APIError"
...
  /v1/me/activity-history/clock-in:
    post:
      tags: [Events]
      summary: Record a clock-in session
      operationId: recordClockIn
      security:
        - UserBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecordClockInRequest"
      responses:
        "201":
          description: Created session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserActivitySession"
        default:
          $ref: "#/components/responses/APIError"
  /v1/me/activity-history/clock-out:
    post:
      tags: [Events]
      summary: Record a clock-out for an open session
      operationId: recordClockOut
      security:
        - UserBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecordClockOutRequest"
      responses:
        "200":
          description: Updated session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserActivitySession"
        default:
          $ref: "#/components/responses/APIError"
...
components:
  securitySchemes:
    BusinessBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Dashboard business user token
    UserBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Mobile end-user token
...
  responses:
    APIError:
      description: Standard API error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIError"
  schemas:
    APIError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
...
    UserEventJoin:
      type: object
      required: [id, event_id, event_title, join_source, joined_at]
      properties:
        id:
          type: string
        event_id:
          type: string
        event_title:
          type: string
        event_description:
          type: string
        event_start_at:
          type: string
          format: date-time
        event_end_at:
          type: string
          format: date-time
        event_status:
          type: string
        join_source:
          type: string
        joined_at:
          type: string
          format: date-time
    ListUserEventJoinsResponse:
      type: object
      required: [joins, total, page, limit, has_more]
      properties:
        joins:
          type: array
          items:
            $ref: "#/components/schemas/UserEventJoin"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        has_more:
          type: boolean
...
    UserActivitySession:
      type: object
      required: [id, user_id, fence_name, clock_in_at, verified, source, created_at]
      properties:
        id:
          type: string
        user_id:
          type: string
        event_id:
          type: string
        event_title:
          type: string
        fence_id:
          type: string
        fence_name:
          type: string
        clock_in_at:
          type: string
          format: date-time
        clock_out_at:
          type: string
          format: date-time
        verified:
          type: boolean
        source:
          type: string
        created_at:
          type: string
          format: date-time
...
    RecordClockInRequest:
      type: object
      properties:
        event_id:
          type: string
        event_title:
          type: string
        fence_id:
          type: string
        fence_name:
          type: string
        source:
          type: string
        lat:
          type: number
        lng:
          type: number
    RecordClockOutRequest:
      type: object
      properties:
        session_id:
          type: string
        event_id:
          type: string
        fence_id:
          type: string
...